refactor: extract shared fetchConversationWithParts helper to deduplicate conversation-fetching logic
